Backend/requirements.txt
requests
pymongo
bcrypt
python-dotenv
argparse
langchain
langchain_huggingface
huggingface_hub
